Se agregan comportamientos a las clases usuario e inmueble

+++ src/trabajoFinal/SitioWeb/PrecioEspecifico.java
@@ -1,29 +1,28 @@
 package trabajoFinal.SitioWeb;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 public class PrecioEspecifico {
-	
-	private int precioPorPeriodo;
+
+	private double precioPorPeriodo;
 	private LocalDate fechaInicial;
 	private LocalDate fechaFinal;
-	
+
 	public PrecioEspecifico(LocalDate fechaInicial, LocalDate fechaFinal, int precio) {
-		
+
 		this.precioPorPeriodo = precio;
 		this.fechaInicial = fechaInicial;
 		this.fechaFinal = fechaFinal;
 	}
-	
-	public int getPrecioPorPeriodo() {
+
+	public double getPrecioPorPeriodo() {
 		return precioPorPeriodo;
 	}
-	
+
 	public LocalDate getFechaInicial() {
 		return fechaInicial;
 	}
-	
+
 	public LocalDate getFechaFinal() {
 		return fechaFinal;
 	}
